@@ -4,7 +4,7 @@ version: 1.4.0
 homepage: https://github.com/Realank/flutter_datetime_picker
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
