@@ -13,24 +13,24 @@ export 'package:flutter_datetime_picker/src/i18n_model.dart';
 
 typedef DateChangedCallback(DateTime time);
 typedef DateCancelledCallback();
-typedef String StringAtIndexCallBack(int index);
+typedef String? StringAtIndexCallBack(int index);
 
 class DatePicker {
   ///
   /// Display date picker bottom sheet.
   ///
-  static Future<DateTime> showDatePicker(
+  static Future<DateTime?> showDatePicker(
     BuildContext context, {
     bool showTitleActions: true,
-    DateTime minTime,
-    DateTime maxTime,
-    DateChangedCallback onChanged,
-    DateChangedCallback onConfirm,
-    DateCancelledCallback onCancel,
+    DateTime? minTime,
+    DateTime? maxTime,
+    DateChangedCallback? onChanged,
+    DateChangedCallback? onConfirm,
+    DateCancelledCallback? onCancel,
     locale: LocaleType.en,
-    DateTime currentTime,
-    DatePickerTheme theme,
-    Widget title,
+    DateTime? currentTime,
+    DatePickerTheme? theme,
+    Widget? title,
   }) async {
     return await Navigator.push(
         context,
@@ -50,17 +50,17 @@ class DatePicker {
   ///
   /// Display time picker bottom sheet.
   ///
-  static Future<DateTime> showTimePicker(
+  static Future<DateTime?> showTimePicker(
     BuildContext context, {
     bool showTitleActions: true,
     bool showSecondsColumn: true,
-    DateChangedCallback onChanged,
-    DateChangedCallback onConfirm,
-    DateCancelledCallback onCancel,
+    DateChangedCallback? onChanged,
+    DateChangedCallback? onConfirm,
+    DateCancelledCallback? onCancel,
     locale: LocaleType.en,
-    DateTime currentTime,
-    DatePickerTheme theme,
-    Widget title,
+    DateTime? currentTime,
+    DatePickerTheme? theme,
+    Widget? title,
   }) async {
     return await Navigator.push(
       context,
@@ -82,16 +82,16 @@ class DatePicker {
   ///
   /// Display time picker bottom sheet with AM/PM.
   ///
-  static Future<DateTime> showTime12hPicker(
+  static Future<DateTime?> showTime12hPicker(
     BuildContext context, {
     bool showTitleActions: true,
-    DateChangedCallback onChanged,
-    DateChangedCallback onConfirm,
-    DateCancelledCallback onCancel,
+    DateChangedCallback? onChanged,
+    DateChangedCallback? onConfirm,
+    DateCancelledCallback? onCancel,
     locale: LocaleType.en,
-    DateTime currentTime,
-    DatePickerTheme theme,
-    Widget title,
+    DateTime? currentTime,
+    DatePickerTheme? theme,
+    Widget? title,
   }) async {
     return await Navigator.push(
         context,
@@ -110,18 +110,18 @@ class DatePicker {
   ///
   /// Display date&time picker bottom sheet.
   ///
-  static Future<DateTime> showDateTimePicker(
+  static Future<DateTime?> showDateTimePicker(
     BuildContext context, {
     bool showTitleActions: true,
-    DateTime minTime,
-    DateTime maxTime,
-    DateChangedCallback onChanged,
-    DateChangedCallback onConfirm,
-    DateCancelledCallback onCancel,
+    DateTime? minTime,
+    DateTime? maxTime,
+    DateChangedCallback? onChanged,
+    DateChangedCallback? onConfirm,
+    DateCancelledCallback? onCancel,
     locale: LocaleType.en,
-    DateTime currentTime,
-    DatePickerTheme theme,
-    Widget title,
+    DateTime? currentTime,
+    DatePickerTheme? theme,
+    Widget? title,
   }) async {
     return await Navigator.push(
         context,
@@ -141,16 +141,16 @@ class DatePicker {
   ///
   /// Display date picker bottom sheet witch custom picker model.
   ///
-  static Future<T> showPicker<T>(
+  static Future<T?> showPicker<T>(
     BuildContext context, {
     bool showTitleActions: true,
-    DateChangedCallback onChanged,
-    DateChangedCallback onConfirm,
-    DateCancelledCallback onCancel,
+    DateChangedCallback? onChanged,
+    DateChangedCallback? onConfirm,
+    DateCancelledCallback? onCancel,
     locale: LocaleType.en,
-    BasePickerModel pickerModel,
-    DatePickerTheme theme,
-    Widget title,
+    BasePickerModel? pickerModel,
+    DatePickerTheme? theme,
+    Widget? title,
   }) async {
     return await Navigator.push(
       context,
@@ -179,20 +179,20 @@ class _DatePickerRoute<T> extends PopupRoute<T> {
     this.barrierLabel,
     this.locale,
     this.title,
-    RouteSettings settings,
+    RouteSettings? settings,
     pickerModel,
   })  : this.pickerModel = pickerModel ?? DatePickerModel(),
         this.theme = theme ?? DatePickerTheme(),
         super(settings: settings);
 
-  final bool showTitleActions;
-  final DateChangedCallback onChanged;
-  final DateChangedCallback onConfirm;
-  final DateCancelledCallback onCancel;
+  final bool? showTitleActions;
+  final DateChangedCallback? onChanged;
+  final DateChangedCallback? onConfirm;
+  final DateCancelledCallback? onCancel;
   final DatePickerTheme theme;
-  final LocaleType locale;
+  final LocaleType? locale;
   final BasePickerModel pickerModel;
-  final Widget title;
+  final Widget? title;
 
   @override
   Duration get transitionDuration => const Duration(milliseconds: 200);
@@ -201,18 +201,18 @@ class _DatePickerRoute<T> extends PopupRoute<T> {
   bool get barrierDismissible => true;
 
   @override
-  final String barrierLabel;
+  final String? barrierLabel;
 
   @override
   Color get barrierColor => Colors.black54;
 
-  AnimationController _animationController;
+  AnimationController? _animationController;
 
   @override
   AnimationController createAnimationController() {
     assert(_animationController == null);
-    _animationController = BottomSheet.createAnimationController(navigator.overlay);
-    return _animationController;
+    _animationController = BottomSheet.createAnimationController(navigator!.overlay!);
+    return _animationController!;
   }
 
   @override
@@ -239,16 +239,16 @@ class _DatePickerRoute<T> extends PopupRoute<T> {
 
 class _DatePickerComponent extends StatefulWidget {
   _DatePickerComponent(
-      {Key key, @required this.route, this.onChanged, this.locale, this.pickerModel, this.title});
+      {Key? key, required this.route, this.onChanged, this.locale, this.pickerModel, this.title});
 
-  final DateChangedCallback onChanged;
+  final DateChangedCallback? onChanged;
 
   final _DatePickerRoute route;
 
-  final LocaleType locale;
+  final LocaleType? locale;
 
-  final BasePickerModel pickerModel;
-  final Widget title;
+  final BasePickerModel? pickerModel;
+  final Widget? title;
 
   @override
   State<StatefulWidget> createState() {
@@ -257,7 +257,7 @@ class _DatePickerComponent extends StatefulWidget {
 }
 
 class _DatePickerState extends State<_DatePickerComponent> {
-  FixedExtentScrollController leftScrollCtrl, middleScrollCtrl, rightScrollCtrl;
+  FixedExtentScrollController? leftScrollCtrl, middleScrollCtrl, rightScrollCtrl;
 
   @override
   void initState() {
@@ -268,11 +268,11 @@ class _DatePickerState extends State<_DatePickerComponent> {
   void refreshScrollOffset() {
 //    print('refreshScrollOffset ${widget.pickerModel.currentRightIndex()}');
     leftScrollCtrl =
-        new FixedExtentScrollController(initialItem: widget.pickerModel.currentLeftIndex());
+        new FixedExtentScrollController(initialItem: widget.pickerModel!.currentLeftIndex()!);
     middleScrollCtrl =
-        new FixedExtentScrollController(initialItem: widget.pickerModel.currentMiddleIndex());
+        new FixedExtentScrollController(initialItem: widget.pickerModel!.currentMiddleIndex()!);
     rightScrollCtrl =
-        new FixedExtentScrollController(initialItem: widget.pickerModel.currentRightIndex());
+        new FixedExtentScrollController(initialItem: widget.pickerModel!.currentRightIndex()!);
   }
 
   @override
@@ -280,12 +280,12 @@ class _DatePickerState extends State<_DatePickerComponent> {
     DatePickerTheme theme = widget.route.theme;
     return GestureDetector(
       child: AnimatedBuilder(
-        animation: widget.route.animation,
-        builder: (BuildContext context, Widget child) {
+        animation: widget.route.animation!,
+        builder: (BuildContext context, Widget? child) {
           final double bottomPadding = MediaQuery.of(context).padding.bottom;
           return ClipRect(
             child: CustomSingleChildLayout(
-              delegate: _BottomPickerLayout(widget.route.animation.value, theme,
+              delegate: _BottomPickerLayout(widget.route.animation!.value, theme,
                   showTitleActions: widget.route.showTitleActions, bottomPadding: bottomPadding),
               child: GestureDetector(
                 child: Material(
@@ -302,13 +302,13 @@ class _DatePickerState extends State<_DatePickerComponent> {
 
   void _notifyDateChanged() {
     if (widget.onChanged != null) {
-      widget.onChanged(widget.pickerModel.finalTime());
+      widget.onChanged!(widget.pickerModel!.finalTime());
     }
   }
 
   Widget _renderPickerView(DatePickerTheme theme) {
     Widget itemView = _renderItemView(theme);
-    if (widget.route.showTitleActions) {
+    if (widget.route.showTitleActions!) {
       return Column(
         children: <Widget>[
           _renderTitleActionsView(theme),
@@ -323,10 +323,10 @@ class _DatePickerState extends State<_DatePickerComponent> {
       ValueKey key,
       DatePickerTheme theme,
       StringAtIndexCallBack stringAtIndexCB,
-      ScrollController scrollController,
+      ScrollController? scrollController,
       int layoutProportion,
       ValueChanged<int> selectedChangedWhenScrolling,
-      ValueChanged<int> selectedChangedWhenScrollEnd) {
+      ValueChanged<int>? selectedChangedWhenScrollEnd) {
     return Expanded(
       flex: layoutProportion,
       child: Container(
@@ -339,7 +339,7 @@ class _DatePickerState extends State<_DatePickerComponent> {
                     selectedChangedWhenScrollEnd != null &&
                     notification is ScrollEndNotification &&
                     notification.metrics is FixedExtentMetrics) {
-                  final FixedExtentMetrics metrics = notification.metrics;
+                  final FixedExtentMetrics metrics = notification.metrics as FixedExtentMetrics;
                   final int currentItemIndex = metrics.itemIndex;
                   selectedChangedWhenScrollEnd(currentItemIndex);
                 }
@@ -348,7 +348,7 @@ class _DatePickerState extends State<_DatePickerComponent> {
               child: CupertinoPicker.builder(
                   key: key,
                   backgroundColor: theme.backgroundColor ?? Colors.white,
-                  scrollController: scrollController,
+                  scrollController: scrollController as FixedExtentScrollController?,
                   itemExtent: theme.itemHeight,
                   onSelectedItemChanged: (int index) {
                     selectedChangedWhenScrolling(index);
@@ -379,14 +379,14 @@ class _DatePickerState extends State<_DatePickerComponent> {
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: <Widget>[
           Container(
-            child: widget.pickerModel.layoutProportions()[0] > 0
+            child: widget.pickerModel!.layoutProportions()[0] > 0
                 ? _renderColumnView(
-                    ValueKey(widget.pickerModel.currentLeftIndex()),
+                    ValueKey(widget.pickerModel!.currentLeftIndex()),
                     theme,
-                    widget.pickerModel.leftStringAtIndex,
+                    widget.pickerModel!.leftStringAtIndex,
                     leftScrollCtrl,
-                    widget.pickerModel.layoutProportions()[0], (index) {
-                    widget.pickerModel.setLeftIndex(index);
+                    widget.pickerModel!.layoutProportions()[0], (index) {
+                    widget.pickerModel!.setLeftIndex(index);
                   }, (index) {
                     setState(() {
                       refreshScrollOffset();
@@ -396,18 +396,18 @@ class _DatePickerState extends State<_DatePickerComponent> {
                 : null,
           ),
           Text(
-            widget.pickerModel.leftDivider(),
+            widget.pickerModel!.leftDivider(),
             style: theme.itemStyle,
           ),
           Container(
-            child: widget.pickerModel.layoutProportions()[1] > 0
+            child: widget.pickerModel!.layoutProportions()[1] > 0
                 ? _renderColumnView(
-                    ValueKey(widget.pickerModel.currentLeftIndex()),
+                    ValueKey(widget.pickerModel!.currentLeftIndex()),
                     theme,
-                    widget.pickerModel.middleStringAtIndex,
+                    widget.pickerModel!.middleStringAtIndex,
                     middleScrollCtrl,
-                    widget.pickerModel.layoutProportions()[1], (index) {
-                    widget.pickerModel.setMiddleIndex(index);
+                    widget.pickerModel!.layoutProportions()[1], (index) {
+                    widget.pickerModel!.setMiddleIndex(index);
                   }, (index) {
                     setState(() {
                       refreshScrollOffset();
@@ -417,19 +417,19 @@ class _DatePickerState extends State<_DatePickerComponent> {
                 : null,
           ),
           Text(
-            widget.pickerModel.rightDivider(),
+            widget.pickerModel!.rightDivider(),
             style: theme.itemStyle,
           ),
           Container(
-            child: widget.pickerModel.layoutProportions()[2] > 0
+            child: widget.pickerModel!.layoutProportions()[2] > 0
                 ? _renderColumnView(
-                    ValueKey(widget.pickerModel.currentMiddleIndex() * 100 +
-                        widget.pickerModel.currentLeftIndex()),
+                    ValueKey(widget.pickerModel!.currentMiddleIndex()! * 100 +
+                        widget.pickerModel!.currentLeftIndex()!),
                     theme,
-                    widget.pickerModel.rightStringAtIndex,
+                    widget.pickerModel!.rightStringAtIndex,
                     rightScrollCtrl,
-                    widget.pickerModel.layoutProportions()[2], (index) {
-                    widget.pickerModel.setRightIndex(index);
+                    widget.pickerModel!.layoutProportions()[2], (index) {
+                    widget.pickerModel!.setRightIndex(index);
                     _notifyDateChanged();
                   }, null)
                 : null,
@@ -441,8 +441,8 @@ class _DatePickerState extends State<_DatePickerComponent> {
 
   // Title View
   Widget _renderTitleActionsView(DatePickerTheme theme) {
-    String done = _localeDone();
-    String cancel = _localeCancel();
+    String? done = _localeDone();
+    String? cancel = _localeCancel();
 
     return Container(
       height: theme.titleHeight,
@@ -464,7 +464,7 @@ class _DatePickerState extends State<_DatePickerComponent> {
               onPressed: () {
                 Navigator.pop(context);
                 if (widget.route.onCancel != null) {
-                  widget.route.onCancel();
+                  widget.route.onCancel!();
                 }
               },
             ),
@@ -480,9 +480,9 @@ class _DatePickerState extends State<_DatePickerComponent> {
                 style: theme.doneStyle,
               ),
               onPressed: () {
-                Navigator.pop(context, widget.pickerModel.finalTime());
+                Navigator.pop(context, widget.pickerModel!.finalTime());
                 if (widget.route.onConfirm != null) {
-                  widget.route.onConfirm(widget.pickerModel.finalTime());
+                  widget.route.onConfirm!(widget.pickerModel!.finalTime());
                 }
               },
             ),
@@ -492,12 +492,12 @@ class _DatePickerState extends State<_DatePickerComponent> {
     );
   }
 
-  String _localeDone() {
-    return i18nObjInLocale(widget.locale)['done'];
+  String? _localeDone() {
+    return i18nObjInLocale(widget.locale)!['done'];
   }
 
-  String _localeCancel() {
-    return i18nObjInLocale(widget.locale)['cancel'];
+  String? _localeCancel() {
+    return i18nObjInLocale(widget.locale)!['cancel'];
   }
 }
 
@@ -506,15 +506,15 @@ class _BottomPickerLayout extends SingleChildLayoutDelegate {
       {this.itemCount, this.showTitleActions, this.bottomPadding = 0});
 
   final double progress;
-  final int itemCount;
-  final bool showTitleActions;
+  final int? itemCount;
+  final bool? showTitleActions;
   final DatePickerTheme theme;
   final double bottomPadding;
 
   @override
   BoxConstraints getConstraintsForChild(BoxConstraints constraints) {
     double maxHeight = theme.containerHeight;
-    if (showTitleActions) {
+    if (showTitleActions!) {
       maxHeight += theme.titleHeight;
     }
 
