@@ -220,14 +220,14 @@ const String am = 'am';
 const String z = 'z';
 const String Z = 'Z';
 
-String formatDate(DateTime date, List<String> formats, LocaleType locale) {
+String? formatDate(DateTime date, List<String> formats, LocaleType locale) {
   if (formats.first == ymdw) {
     final now = DateTime.now();
     if (date.year == now.year &&
         date.month == now.month &&
         date.day == now.day) {
       //today
-      return i18nObjInLocale(locale)['today'];
+      return i18nObjInLocale(locale)!['today'];
     } else if (date.year == now.year) {
       if (locale == LocaleType.zh) {
         return formatDate(date, [mm, '月', dd, '日 ', D], locale);
@@ -283,10 +283,10 @@ String formatDate(DateTime date, List<String> formats, LocaleType locale) {
     } else if (format == m) {
       sb.write(date.month);
     } else if (format == MM) {
-      String monthLong = i18nObjInLocale(locale)['monthLong'][date.month - 1];
+      String? monthLong = i18nObjInLocale(locale)!['monthLong'][date.month - 1];
       sb.write(monthLong);
     } else if (format == M) {
-      String monthShort = i18nObjInLocale(locale)['monthShort'][date.month - 1];
+      String? monthShort = i18nObjInLocale(locale)!['monthShort'][date.month - 1];
       sb.write(monthShort);
     } else if (format == dd) {
       sb.write(digits(date.day, 2));
@@ -299,7 +299,7 @@ String formatDate(DateTime date, List<String> formats, LocaleType locale) {
     } else if (format == WW) {
       sb.write(digits((dayInYear(date) + 7) ~/ 7, 2));
     } else if (format == D) {
-      String day = i18nObjInLocale(locale)['day'][date.weekday - 1];
+      String? day = i18nObjInLocale(locale)!['day'][date.weekday - 1];
       if (locale == LocaleType.ko) {
         day = "($day)";
       }
@@ -314,8 +314,8 @@ String formatDate(DateTime date, List<String> formats, LocaleType locale) {
       sb.write(date.hour % 12);
     } else if (format == am) {
       sb.write(date.hour < 12
-          ? i18nObjInLocale(locale)['am']
-          : i18nObjInLocale(locale)['pm']);
+          ? i18nObjInLocale(locale)!['am']
+          : i18nObjInLocale(locale)!['pm']);
     } else if (format == nn) {
       sb.write(digits(date.minute, 2));
     } else if (format == n) {
